+++ addOns/pscanrules/src/main/java/org/zaproxy/zap/extension/pscanrules/ContentSecurityPolicyScanRule.java
@@ -37,7 +37,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import net.htmlparser.jericho.Source;
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.Constant;
@@ -120,26 +120,22 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
         // 4.0+, and IE 10+ (in a limited fashion)
         List<String> xcspOptions = msg.getResponseHeader().getHeaderValues(HTTP_HEADER_XCSP);
         if (!xcspOptions.isEmpty()) {
-            raiseAlert(
-                    Constant.messages.getString(MESSAGE_PREFIX + "xcsp.name"),
-                    Constant.messages.getString(MESSAGE_PREFIX + "xcsp.otherinfo"),
-                    getHeaderField(msg, HTTP_HEADER_XCSP).get(0),
-                    cspHeaderFound ? Alert.RISK_INFO : Alert.RISK_LOW,
-                    xcspOptions.get(0),
-                    "1");
+            buildXcspAlert(
+                            cspHeaderFound ? Alert.RISK_INFO : Alert.RISK_LOW,
+                            getHeaderField(msg, HTTP_HEADER_XCSP).get(0),
+                            xcspOptions.get(0))
+                    .raise();
         }
 
         // X-WebKit-CSP is supported by Chrome 14+, and Safari 6+
         List<String> xwkcspOptions =
                 msg.getResponseHeader().getHeaderValues(HTTP_HEADER_WEBKIT_CSP);
         if (!xwkcspOptions.isEmpty()) {
-            raiseAlert(
-                    Constant.messages.getString(MESSAGE_PREFIX + "xwkcsp.name"),
-                    Constant.messages.getString(MESSAGE_PREFIX + "xwkcsp.otherinfo"),
-                    getHeaderField(msg, HTTP_HEADER_WEBKIT_CSP).get(0),
-                    cspHeaderFound ? Alert.RISK_INFO : Alert.RISK_LOW,
-                    xwkcspOptions.get(0),
-                    "2");
+            buildWebkitCspAlert(
+                            cspHeaderFound ? Alert.RISK_INFO : Alert.RISK_LOW,
+                            getHeaderField(msg, HTTP_HEADER_WEBKIT_CSP).get(0),
+                            xwkcspOptions.get(0))
+                    .raise();
         }
 
         if (cspHeaderFound) {
@@ -168,13 +164,12 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
                     } else {
                         noticesRisk = Alert.RISK_INFO;
                     }
-                    raiseAlert(
-                            Constant.messages.getString(MESSAGE_PREFIX + "notices.name"),
-                            cspNoticesString,
-                            "",
-                            noticesRisk,
-                            csp,
-                            "3");
+                    buildNoticesAlert(
+                                    noticesRisk,
+                                    getHeaderField(msg, HTTP_HEADER_CSP).get(0),
+                                    csp,
+                                    cspNoticesString)
+                            .raise();
                 }
 
                 List<String> allowedWildcardSources = getAllowedWildcardSources(csp);
@@ -194,65 +189,34 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
                                         "pscanrules.csp.otherinfo.extended",
                                         String.join(", ", allowedDirectivesWithoutFallback));
                     }
-                    raiseAlert(
-                            Constant.messages.getString(MESSAGE_PREFIX + "wildcard.name"),
-                            wildcardSrcOtherInfo,
-                            "",
-                            Alert.RISK_MEDIUM,
-                            csp,
-                            "4");
+                    buildWildcardAlert(
+                                    getHeaderField(msg, HTTP_HEADER_CSP).get(0),
+                                    csp,
+                                    wildcardSrcOtherInfo)
+                            .raise();
                 }
 
                 PolicyInOrigin p = new PolicyInOrigin(policy, URI.parseURI(RAND_FQDN).orElse(null));
                 if (p.allowsUnsafeInlineScript()) {
-                    raiseAlert(
-                            Constant.messages.getString(MESSAGE_PREFIX + "scriptsrc.unsafe.name"),
-                            Constant.messages.getString(
-                                    MESSAGE_PREFIX + "scriptsrc.unsafe.otherinfo"),
-                            "",
-                            Alert.RISK_MEDIUM,
-                            csp,
-                            "5");
+                    buildScriptUnsafeInlineAlert(getHeaderField(msg, HTTP_HEADER_CSP).get(0), csp)
+                            .raise();
                 }
 
                 if (p.allowsUnsafeInlineStyle()) {
-                    raiseAlert(
-                            Constant.messages.getString(MESSAGE_PREFIX + "stylesrc.unsafe.name"),
-                            Constant.messages.getString(
-                                    MESSAGE_PREFIX + "stylesrc.unsafe.otherinfo"),
-                            "",
-                            Alert.RISK_MEDIUM,
-                            csp,
-                            "6");
+                    buildStyleUnsafeInlineAlert(getHeaderField(msg, HTTP_HEADER_CSP).get(0), csp)
+                            .raise();
                 }
 
                 if (allowsUnsafeHashes(policy, FetchDirectiveKind.ScriptSrc)) {
-                    raiseAlert(
-                            Constant.messages.getString(
-                                    MESSAGE_PREFIX + "scriptsrc.unsafe.hashes.name"),
-                            Constant.messages.getString(
-                                    MESSAGE_PREFIX + "scriptsrc.unsafe.hashes.otherinfo"),
-                            "",
-                            Alert.RISK_MEDIUM,
-                            csp,
-                            Constant.messages.getString(
-                                    MESSAGE_PREFIX + "scriptsrc.unsafe.hashes.refs"),
-                            "7");
+                    buildScriptUnsafeHashAlert(getHeaderField(msg, HTTP_HEADER_CSP).get(0), csp)
+                            .raise();
                     ;
                 }
 
                 if (allowsUnsafeHashes(policy, FetchDirectiveKind.StyleSrc)) {
-                    raiseAlert(
-                            Constant.messages.getString(
-                                    MESSAGE_PREFIX + "stylesrc.unsafe.hashes.name"),
-                            Constant.messages.getString(
-                                    MESSAGE_PREFIX + "stylesrc.unsafe.hashes.otherinfo"),
-                            "",
-                            Alert.RISK_MEDIUM,
-                            csp,
-                            Constant.messages.getString(
-                                    MESSAGE_PREFIX + "stylesrc.unsafe.hashes.refs"),
-                            "8");
+                    buildStyleUnsafeHashAlert(getHeaderField(msg, HTTP_HEADER_CSP).get(0), csp)
+                            .raise();
+                    ;
                 }
             }
         }
@@ -450,40 +414,139 @@ public class ContentSecurityPolicyScanRule extends PluginPassiveScanner {
         return 15; // WASC-15: Application Misconfiguration
     }
 
-    private void raiseAlert(
-            String name,
-            String otherInfo,
-            String param,
-            int risk,
-            String evidence,
-            String alertRef) {
-        raiseAlert(name, otherInfo, param, risk, evidence, getReference(), alertRef);
-    }
-
-    private void raiseAlert(
-            String name,
-            String otherInfo,
-            String param,
-            int risk,
-            String evidence,
-            String refs,
-            String alertRef) {
+    private AlertBuilder getBuilder(String name, String alertRef) {
         String alertName = StringUtils.isEmpty(name) ? getName() : getName() + ": " + name;
-
-        newAlert()
+        return newAlert()
                 .setName(alertName)
-                .setRisk(risk)
-                .setConfidence(Alert.CONFIDENCE_MEDIUM)
+                .setConfidence(Alert.CONFIDENCE_HIGH)
                 .setDescription(Constant.messages.getString(MESSAGE_PREFIX + "desc"))
-                .setOtherInfo(otherInfo)
-                .setParam(param)
                 .setSolution(getSolution())
-                .setReference(refs)
-                .setEvidence(evidence)
-                .setAlertRef(PLUGIN_ID + "-" + alertRef)
+                .setReference(getReference())
                 .setCweId(getCweId())
                 .setWascId(getWascId())
-                .raise();
+                .setAlertRef(PLUGIN_ID + "-" + alertRef);
+    }
+
+    private AlertBuilder buildXcspAlert(int risk, String param, String evidence) {
+        return getBuilder(Constant.messages.getString(MESSAGE_PREFIX + "xcsp.name"), "1")
+                .setRisk(risk)
+                .setParam(param)
+                .setEvidence(evidence)
+                .setOtherInfo(Constant.messages.getString(MESSAGE_PREFIX + "xcsp.otherinfo"));
+    }
+
+    private AlertBuilder buildWebkitCspAlert(int risk, String param, String evidence) {
+        return getBuilder(Constant.messages.getString(MESSAGE_PREFIX + "xwkcsp.name"), "2")
+                .setRisk(risk)
+                .setParam(param)
+                .setEvidence(evidence)
+                .setOtherInfo(Constant.messages.getString(MESSAGE_PREFIX + "xwkcsp.otherinfo"));
+    }
+
+    private AlertBuilder buildNoticesAlert(
+            int risk, String param, String evidence, String otherinfo) {
+        return getBuilder(Constant.messages.getString(MESSAGE_PREFIX + "notices.name"), "3")
+                .setRisk(risk)
+                .setParam(param)
+                .setEvidence(evidence)
+                .setOtherInfo(otherinfo);
+    }
+
+    private AlertBuilder buildWildcardAlert(String param, String evidence, String otherinfo) {
+        return getBuilder(Constant.messages.getString(MESSAGE_PREFIX + "wildcard.name"), "4")
+                .setRisk(Alert.RISK_MEDIUM)
+                .setParam(param)
+                .setEvidence(evidence)
+                .setOtherInfo(otherinfo);
+    }
+
+    private AlertBuilder buildScriptUnsafeInlineAlert(String param, String evidence) {
+        return getBuilder(
+                        Constant.messages.getString(MESSAGE_PREFIX + "scriptsrc.unsafe.name"), "5")
+                .setRisk(Alert.RISK_MEDIUM)
+                .setParam(param)
+                .setEvidence(evidence)
+                .setOtherInfo(
+                        Constant.messages.getString(MESSAGE_PREFIX + "scriptsrc.unsafe.otherinfo"));
+    }
+
+    private AlertBuilder buildStyleUnsafeInlineAlert(String param, String evidence) {
+        return getBuilder(Constant.messages.getString(MESSAGE_PREFIX + "stylesrc.unsafe.name"), "6")
+                .setRisk(Alert.RISK_MEDIUM)
+                .setParam(param)
+                .setEvidence(evidence)
+                .setOtherInfo(
+                        Constant.messages.getString(MESSAGE_PREFIX + "stylesrc.unsafe.otherinfo"));
+    }
+
+    private AlertBuilder buildScriptUnsafeHashAlert(String param, String evidence) {
+        return getBuilder(
+                        Constant.messages.getString(
+                                MESSAGE_PREFIX + "scriptsrc.unsafe.hashes.name"),
+                        "7")
+                .setRisk(Alert.RISK_MEDIUM)
+                .setParam(param)
+                .setEvidence(evidence)
+                .setOtherInfo(
+                        Constant.messages.getString(
+                                MESSAGE_PREFIX + "scriptsrc.unsafe.hashes.otherinfo"))
+                .setReference(
+                        Constant.messages.getString(
+                                MESSAGE_PREFIX + "scriptsrc.unsafe.hashes.refs"));
+    }
+
+    private AlertBuilder buildStyleUnsafeHashAlert(String param, String evidence) {
+        return getBuilder(
+                        Constant.messages.getString(MESSAGE_PREFIX + "stylesrc.unsafe.hashes.name"),
+                        "8")
+                .setRisk(Alert.RISK_MEDIUM)
+                .setParam(param)
+                .setEvidence(evidence)
+                .setOtherInfo(
+                        Constant.messages.getString(
+                                MESSAGE_PREFIX + "stylesrc.unsafe.hashes.otherinfo"))
+                .setReference(
+                        Constant.messages.getString(
+                                MESSAGE_PREFIX + "stylesrc.unsafe.hashes.refs"));
+    }
+
+    @Override
+    public List<Alert> getExampleAlerts() {
+        List<Alert> alerts = new ArrayList<>();
+        alerts.add(buildXcspAlert(Alert.RISK_LOW, "default-src 'self'", HTTP_HEADER_XCSP).build());
+        alerts.add(
+                buildWebkitCspAlert(Alert.RISK_LOW, "default-src 'self'", HTTP_HEADER_WEBKIT_CSP)
+                        .build());
+        alerts.add(
+                buildNoticesAlert(
+                                Alert.RISK_LOW,
+                                HTTP_HEADER_CSP,
+                                "default-src none; report-to csp-endpoint ",
+                                "Warnings:\\nThis host name is unusual, and likely meant to be a keyword that is missing the required quotes: 'none'.")
+                        .build());
+        alerts.add(
+                buildWildcardAlert(
+                                HTTP_HEADER_CSP,
+                                "connect-src *; default-src 'self'; form-action 'none'; frame-ancestors 'self'",
+                                Constant.messages.getString(
+                                        MESSAGE_PREFIX + "wildcard.otherinfo", "connect-src"))
+                        .build());
+        alerts.add(
+                buildScriptUnsafeInlineAlert(HTTP_HEADER_CSP, "script-src 'unsafe-inline'")
+                        .build());
+        alerts.add(
+                buildStyleUnsafeInlineAlert(HTTP_HEADER_CSP, "style-src 'unsafe-inline'").build());
+        alerts.add(
+                buildScriptUnsafeHashAlert(
+                                HTTP_HEADER_CSP,
+                                "default-src 'self'; script-src 'unsafe-hashes' 'sha256-jzgBGA4UWFFmpOBq0JpdsySukE1FrEN5bUpoK8Z29fY='")
+                        .build());
+        alerts.add(
+                buildStyleUnsafeHashAlert(
+                                HTTP_HEADER_CSP,
+                                "default-src 'self'; style-src 'unsafe-hashes' 'sha256-xyz4zkCjuC3lZcD2UmnqDG0vurmq12W/XKM5Vd0+MlQ='")
+                        .build());
+        return alerts;
     }
 
     static class PolicyError {